refactor(SPV-1058): provide method for setting up engine without server.
